UnifiedDashboard/script/boost_github_script/requirements.txt
SQLAlchemy>=2.0,<3.0
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
psycopg2-binary>=2.9,<3.0


